+++ pysad/models/integrations/reference_window_model.py
@@ -9,8 +9,8 @@ class ReferenceWindowModel(PYODModel):
             model_cls (class): The model class to be instantiated.
             window_size (int): The size of each window.
             sliding_size (int): The sliding length of the windows.
-            initial_X (np.float array of shape (num_initial_instances, num_features)): Initial instances to fit.
-            initial_y (np.int array of shape (num_initial_instances,)): Initial window's ground truth labels. Used if not None. Needs to be `None` for the unsupervised `model_cls` models. (Default=None).
+            initial_X (np.float64 array of shape (num_initial_instances, num_features)): Initial instances to fit.
+            initial_y (np.int32 array of shape (num_initial_instances,)): Initial window's ground truth labels. Used if not None. Needs to be `None` for the unsupervised `model_cls` models. (Default=None).
             **kwargs (Keyword arguments): Keyword arguments that is passed to the `model_cls`.
     """
 
@@ -52,7 +52,7 @@ class ReferenceWindowModel(PYODModel):
         """Fits the model to next instance.
 
         Args:
-            X (np.float array of shape (num_features,)): The instance to fit.
+            X (np.float64 array of shape (num_features,)): The instance to fit.
             y (int): The label of the instance (Optional for unsupervised models, default=None).
 
         Returns:
@@ -97,7 +97,7 @@ class ReferenceWindowModel(PYODModel):
         """Scores the anomalousness of the next instance.
 
         Args:
-            X (np.float array of shape (num_features,)): The instance to score. Higher scores represent more anomalous instances whereas lower scores correspond to more normal instances.
+            X (np.float64 array of shape (num_features,)): The instance to score. Higher scores represent more anomalous instances whereas lower scores correspond to more normal instances.
 
         Returns:
             float: The anomalousness score of the input instance.
